+++ src/Mu/Modelling/Services/ReflectionTransform.cs
@@ -3,7 +3,6 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Text;
 using Ardalis.GuardClauses;
 using Mu.Modelling.Behavior;
 using Mu.Modelling.State;