+++ src/Muify.Tests/Service/FeatureGrpcServiceVisitorTests/WhenObserveIsCalled.cs
@@ -1,0 +1,11 @@
+﻿namespace Muify.Service.FeatureGrpcServiceVisitorTests;
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+public sealed class WhenObserveIsCalled
+{
+    // Do we keep the Mediator or do we just use the ServiceHandler directly
+    // What do we do about distributed tracing?
+}